requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
django-filter>=23.1
python-dotenv>=1.0
django-cors-headers>=4.0
psycopg2-binary>=2.9  # optional, for Postgres production
